Got manual data to be added to the database
Right now its just adding "Tokyo" and "Japan" into the collection. Also exported the function and imported it in map.tsx to call the function and see if it can be used there as well.
Next steps would be to create a sign-up form and store their username, email, and password as variables and pass it through the addData function (can be renamed)